Don't re-add moved punctuation where quotes end with it already
